+++ core-ng/src/main/java/core/framework/api/module/MessageHandlerConfig.java
@@ -6,7 +6,7 @@ import core.framework.api.queue.MessageHandler;
  * @author neo
  */
 public interface MessageHandlerConfig {
-    MessageHandlerConfig maxConcurrentHandlers(int maxConcurrentHandlers);
-
     <T> MessageHandlerConfig handle(Class<T> messageClass, MessageHandler<T> handler);
+
+    MessageHandlerConfig maxConcurrentHandlers(int maxConcurrentHandlers);
 }