+++ core-ng/src/main/java/core/framework/impl/module/QueueManager.java
@@ -1,7 +1,11 @@
 package core.framework.impl.module;
 
+import core.framework.api.module.MessageHandlerConfig;
+import core.framework.api.util.Maps;
 import core.framework.impl.queue.MessageValidator;
 import core.framework.impl.queue.RabbitMQ;
+
+import java.util.Map;
 
 /**
  * @author neo
@@ -9,6 +13,14 @@ import core.framework.impl.queue.RabbitMQ;
 public class QueueManager {
     public RabbitMQ rabbitMQ;
     private MessageValidator validator;
+    private Map<String, MessageHandlerConfig> listeners = Maps.newHashMap();
+
+    public Map<String, MessageHandlerConfig> listeners() {
+        if (listeners == null) {
+            listeners = Maps.newHashMap();
+        }
+        return listeners;
+    }
 
     public MessageValidator validator() {
         if (validator == null) {