+++ core-ng/src/main/java/core/framework/impl/queue/RabbitMQListener.java
@@ -19,43 +19,50 @@ import org.slf4j.LoggerFactory;
 
 import java.time.Duration;
 import java.util.Map;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
 import java.util.concurrent.Semaphore;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 /**
  * @author neo
  */
-public class RabbitMQListener implements Runnable, MessageHandlerConfig {
+public class RabbitMQListener implements MessageHandlerConfig {
     static final String HEADER_TRACE = "trace";
     static final String HEADER_CLIENT_IP = "clientIP";
 
     private final Logger logger = LoggerFactory.getLogger(RabbitMQListener.class);
 
-    private final ExecutorService listenerExecutor = Executors.newSingleThreadExecutor();
-    private final Executor executor;
-    private final RabbitMQ rabbitMQ;
-    private final LogManager logManager;
-    private final String queue;
-
+    private final AtomicBoolean shutdown = new AtomicBoolean(false);
+    private final Thread listenerThread;
     private int maxConcurrentHandlers = 10;
     private Semaphore semaphore;
+
+    private final String queue;
+    private final Executor executor;
+    private final LogManager logManager;
     private final MessageValidator validator;
     private final Map<String, MessageHandler> handlers = Maps.newHashMap();
     private final Map<String, Class> messageClasses = Maps.newHashMap();
 
     public RabbitMQListener(RabbitMQ rabbitMQ, String queue, Executor executor, MessageValidator validator, LogManager logManager) {
         this.executor = executor;
-        this.rabbitMQ = rabbitMQ;
         this.queue = queue;
         this.validator = validator;
         this.logManager = logManager;
-    }
 
-    @Override
-    public MessageHandlerConfig maxConcurrentHandlers(int maxConcurrentHandlers) {
-        this.maxConcurrentHandlers = maxConcurrentHandlers;
-        return this;
+        listenerThread = new Thread(() -> {
+            logger.info("rabbitMQ message listener started, queue={}", queue);
+            while (!shutdown.get()) {
+                try (RabbitMQConsumer consumer = rabbitMQ.consumer(queue, maxConcurrentHandlers)) {
+                    pullMessages(consumer);
+                } catch (ShutdownSignalException | InterruptedException e) {
+                    // pass thru if it's shutdown
+                } catch (Throwable e) {
+                    logger.error("failed to pull message, retry in 30 seconds", e);
+                    Threads.sleepRoughly(Duration.ofSeconds(30));
+                }
+            }
+        });
+        listenerThread.setName("rabbitMQ-listener-" + queue);
     }
 
     @Override
@@ -71,26 +78,15 @@ public class RabbitMQListener implements Runnable, MessageHandlerConfig {
     }
 
     @Override
-    public void run() {
-        Thread.currentThread().setName("rabbitMQ-listener-" + Thread.currentThread().getId());
-        logger.info("rabbitMQ message listener started, queue={}", queue);
-        semaphore = new Semaphore(maxConcurrentHandlers, false);
-        while (!listenerExecutor.isShutdown()) {
-            try (RabbitMQConsumer consumer = rabbitMQ.consumer(queue, maxConcurrentHandlers)) {
-                pullMessage(consumer);
-            } catch (ShutdownSignalException e) {
-                // pass thru if it's shutdown
-            } catch (Throwable e) {
-                logger.error("failed to pull message, retry in 30 seconds", e);
-                Threads.sleepRoughly(Duration.ofSeconds(30));
-            }
-        }
+    public MessageHandlerConfig maxConcurrentHandlers(int maxConcurrentHandlers) {
+        this.maxConcurrentHandlers = maxConcurrentHandlers;
+        return this;
     }
 
-    private void pullMessage(RabbitMQConsumer consumer) {
-        while (!listenerExecutor.isShutdown()) {
+    private void pullMessages(RabbitMQConsumer consumer) throws InterruptedException {
+        while (!shutdown.get()) {
             QueueingConsumer.Delivery delivery = consumer.nextDelivery();
-            semaphore.acquireUninterruptibly(); // acquire permit right before submit, to avoid permit failing to release back due to exception in between
+            semaphore.acquire(); // acquire permit right before submit, to avoid permit failing to release back due to exception in between
             executor.submit(() -> {
                 try {
                     process(delivery);
@@ -104,12 +100,14 @@ public class RabbitMQListener implements Runnable, MessageHandlerConfig {
     }
 
     public void start() {
-        listenerExecutor.submit(this);
+        semaphore = new Semaphore(maxConcurrentHandlers, false);
+        listenerThread.start();
     }
 
     public void shutdown() {
         logger.info("shutdown rabbitMQ message listener, queue={}", queue);
-        listenerExecutor.shutdown();
+        shutdown.set(true);
+        listenerThread.interrupt();
     }
 
     private <T> void process(QueueingConsumer.Delivery delivery) throws Exception {